+++ BlockLayer.h
@@ -15,7 +15,7 @@ class DenseLayerComposite {
 
         void initialize(int in_channels);
 
-        Tensor<double, 4> forward(const Tensor<double, 4> &input);
+        Tensor<float, 4> forward(const Tensor<float, 4> &input);
 };
 
 #endif
