+++ Pooling.cpp
@@ -9,27 +9,27 @@ PoolingLayer::PoolingLayer(int input_chan, int output_chan, PoolingType type, in
     out_channels = output_chan;
 }
 
-Tensor<double, 4> PoolingLayer::forward(const Tensor<double, 4> &input) {
+Tensor<float, 4> PoolingLayer::forward(const Tensor<float, 4> &input) {
     int batch_size = input.dimension(0);
     int channels = input.dimension(1);
     int height = input.dimension(2);
     int width = input.dimension(3);
 
     if (mode == GLOBAL) {
-        Tensor<double, 4> output(batch_size, channels, 1, 1);
+        Tensor<float, 4> output(batch_size, channels, 1, 1);
         output.setZero();
 
         for (int b = 0; b < batch_size; b++) {
             for (int c = 0; c < channels; c++) {
-                Tensor<double, 2> feature_map = input.chip(b, 0).chip(c, 0);
+                Tensor<float, 2> feature_map = input.chip(b, 0).chip(c, 0);
                 feature_map.maximum();
                 if (type == MAX) {
-                    Tensor<double, 0> max_val_scl = feature_map.maximum();
-                    double max_val = max_val_scl(0);
+                    Tensor<float, 0> max_val_scl = feature_map.maximum();
+                    float max_val = max_val_scl(0);
                     output(b, c, 0, 0) = max_val;
                 } else {
-                    Tensor<double, 0> mean_val_scl = feature_map.mean();
-                    double mean_val = mean_val_scl(0);
+                    Tensor<float, 0> mean_val_scl = feature_map.mean();
+                    float mean_val = mean_val_scl(0);
                     output(b, c, 0, 0) = mean_val;
                 }
                     
@@ -40,12 +40,12 @@ Tensor<double, 4> PoolingLayer::forward(const Tensor<double, 4> &input) {
         int output_height = (height - pool_size) / stride + 1;
         int output_width = (width - pool_size) / stride + 1;
 
-        Tensor<double, 4> output(batch_size, channels, output_height, output_width);
+        Tensor<float, 4> output(batch_size, channels, output_height, output_width);
         output.setZero();
 
         for (int b = 0; b < batch_size; b++) {
             for (int c = 0; c < channels; c++ ) {
-                Tensor<double, 2> feature_map = input.chip(b, 0).chip(c, 0);
+                Tensor<float, 2> feature_map = input.chip(b, 0).chip(c, 0);
 
                 for (int i = 0; i < output_height; i++) {
                     for (int j = 0; j < output_width; j++) {
@@ -61,12 +61,12 @@ Tensor<double, 4> PoolingLayer::forward(const Tensor<double, 4> &input) {
                         );
 
                         if (type == MAX) {
-                            Tensor<double, 0> max_val_scl = slice.maximum();
-                            double max_val = max_val_scl(0);
+                            Tensor<float, 0> max_val_scl = slice.maximum();
+                            float max_val = max_val_scl(0);
                             output(b, c, i, j) = max_val;
                         } else {
-                            Tensor<double, 0> mean_val_scl = slice.mean();
-                            double mean_val = mean_val_scl(0);
+                            Tensor<float, 0> mean_val_scl = slice.mean();
+                            float mean_val = mean_val_scl(0);
                             output(b, c, i, j) = mean_val;
                         }
